Initial Concept for a value class
The class "Value" could be used to replace the "double" type for things defined by numbers where the numbers are not necessarily fully defined. Would allow us to set either min and/or max limits to the value (which could be enabled or disabled independently). 

Am thinking that mathmatical operations on these objects could be defined in a way that they would use the current value as a double, to allow for conventionally written logic to use these objects with minimal modification.